local-server/requirements.txt
opencv-python
requests
python-dotenv
websockets
cryptography
pygame